move group_norm, data_norm to static
