+++ python/paddle/fluid/layers/nn.py
@@ -73,7 +73,6 @@ __all__ = [
     'pool2d',
     'pool3d',
     'batch_norm',
-    'data_norm',
     'reduce_mean',
     'reduce_all',
     'reduce_any',
@@ -87,7 +86,6 @@ __all__ = [
     'row_conv',
     'multiplex',
     'layer_norm',
-    'group_norm',
     'spectral_norm',
     'smooth_l1',
     'one_hot',
@@ -2474,210 +2472,6 @@ def batch_norm(
     return helper.append_activation(batch_norm_out)
 
 
-@static_only
-def data_norm(
-    input,
-    act=None,
-    epsilon=1e-05,
-    param_attr=None,
-    data_layout='NCHW',
-    in_place=False,
-    name=None,
-    moving_mean_name=None,
-    moving_variance_name=None,
-    do_model_average_for_mean_and_var=True,
-    slot_dim=-1,
-    sync_stats=False,
-    summary_decay_rate=0.9999999,
-    enable_scale_and_shift=False,
-):
-    r"""
-    :api_attr: Static Graph
-
-    **Data Normalization Layer**
-
-    This op can be used as a normalizer function for conv2d and fully_connected operations.
-    The required data format for this layer is one of the following:
-
-    1. NHWC `[batch, in_height, in_width, in_channels]`
-
-    2. NCHW `[batch, in_channels, in_height, in_width]`
-
-    :math:`input` is the input features over a mini-batch.
-
-    ..  math::
-
-        \\mu_{\\beta} &\\gets \\frac{1}{m} \\sum_{i=1}^{m} x_i \\qquad &//\\
-        \ mini-batch\ mean \\\\
-        \\sigma_{\\beta}^{2} &\\gets \\frac{1}{m} \\sum_{i=1}^{m}(x_i - \\
-        \\mu_{\\beta})^2 \\qquad &//\ mini-batch\ variance \\\\
-        \\hat{x_i} &\\gets \\frac{x_i - \\mu_\\beta} {\\sqrt{\\
-        \\sigma_{\\beta}^{2} + \\epsilon}} \\qquad &//\ normalize \\\\
-        y_i &\\gets \\gamma \\hat{x_i} + \\beta \\qquad &//\ scale\ and\ shift
-
-    Args:
-        input(Tensor): The input Tensor.
-        act(string, Default None): Activation type, linear|relu|prelu|...
-        epsilon(float, Default 1e-05):
-        param_attr(ParamAttr): The parameter attribute for Parameter `scale`.
-        data_layout (str, optional): Specify the data format of the input, and the data format of the output
-            will be consistent with that of the input. An optional string from: `"NCHW"`, `"NHWC"`.
-            The default is `"NCHW"`. When it is `"NCHW"`, the data is stored in the order of:
-            `[batch_size, input_channels, input_height, input_width]`.
-        in_place(bool, Default False): Make the input and output of batch norm reuse memory.
-        name(string, Default None): A name for this layer(optional). If set None, the layer
-            will be named automatically.
-        moving_mean_name(string, Default None): The name of moving_mean which store the global Mean.
-        moving_variance_name(string, Default None): The name of the moving_variance which store the global Variance.
-        do_model_average_for_mean_and_var(bool, Default True): Whether parameter mean and variance
-            should do model average when model average is enabled.
-        slot_dim(int): The embedding dimension of one slot. Slot is a set of one specific feature. In pslib mode, we
-            distinguish feature ids by slot and pull their embeddings from parameter server (pslib). The first
-            place of the embedding is the historical show number (occurence time of this feature id with a label 0).
-            If the input of this op is concated by slot-wise embeddings, and the show number is zero when this slot
-            is new or empty, the normalization result may be impractical. To avoid this, we add slot_dim to locate
-            the show number and judge if the show number is zero. If so, we choose to skip normalization on this
-            embedding.
-        sync_stats(bool, Default False): When running with multiple GPU cards, using allreduce to sync the
-            summary messages.
-        summary_decay_rate(float, Default 0.9999999): The decay rate when updating summary.
-        enable_scale_and_shift(bool, Default False): do scale&shift after normalization.
-
-    Returns:
-        Tensor: A tensor which is the result after applying data normalization on the input.
-
-    Examples:
-
-        .. code-block:: python
-
-            import paddle
-            paddle.enable_static()
-
-            x = paddle.randn(shape=[32,100])
-            hidden2 = paddle.static.nn.data_norm(input=x)
-    """
-    helper = LayerHelper('data_norm', **locals())
-    dtype = helper.input_dtype()
-
-    input_shape = input.shape
-    if data_layout == 'NCHW':
-        channel_num = input_shape[1]
-    else:
-        if data_layout == 'NHWC':
-            channel_num = input_shape[-1]
-        else:
-            raise ValueError("unsupported data layout:" + data_layout)
-
-    param_shape = [channel_num]
-
-    batch_size_default = 1e4
-    batch_sum_default = 0.0
-    batch_square_sum_default = 1e4
-    scale_w_default = 1.0
-    bias_default = 0.0
-
-    if param_attr and isinstance(param_attr, dict):
-        batch_size_default = param_attr.get("batch_size", 1e4)
-        batch_sum_default = param_attr.get("batch_sum", 0.0)
-        batch_square_sum_default = param_attr.get("batch_square", 1e4)
-    if enable_scale_and_shift:
-        scale_w_default = param_attr.get("scale_w", 1.0)
-        bias_default = param_attr.get("bias", 0.0)
-
-    # create scale and shift(bias) when enable_scale_and_shift is True
-    if name is None:
-        name = "dn"
-    if enable_scale_and_shift:
-        scale_w = helper.create_parameter(
-            attr=ParamAttr(
-                name=name + '.scale_w',
-                initializer=Constant(value=float(scale_w_default)),
-                trainable=True,
-            ),
-            shape=param_shape,
-            dtype=input.dtype,
-        )
-        bias = helper.create_parameter(
-            attr=ParamAttr(
-                name=name + '.bias',
-                initializer=Constant(value=float(bias_default)),
-                trainable=True,
-            ),
-            shape=param_shape,
-            dtype=input.dtype,
-        )
-    # create parameter
-    batch_size = helper.create_parameter(
-        attr=ParamAttr(
-            name=name + '.batch_size',
-            initializer=Constant(value=float(batch_size_default)),
-            trainable=True,
-        ),
-        shape=param_shape,
-        dtype=input.dtype,
-    )
-
-    batch_sum = helper.create_parameter(
-        attr=ParamAttr(
-            name=name + '.batch_sum',
-            initializer=Constant(value=float(batch_sum_default)),
-            trainable=True,
-        ),
-        shape=param_shape,
-        dtype=input.dtype,
-    )
-
-    batch_square_sum = helper.create_parameter(
-        attr=ParamAttr(
-            name=name + '.batch_square_sum',
-            initializer=Constant(value=float(batch_square_sum_default)),
-            trainable=True,
-        ),
-        shape=param_shape,
-        dtype=input.dtype,
-    )
-
-    means = helper.create_variable(dtype=dtype, stop_gradient=True)
-    scales = helper.create_variable(dtype=dtype, stop_gradient=True)
-
-    data_norm_out = input if in_place else helper.create_variable(dtype=dtype)
-
-    inputs = {
-        "X": input,
-        "BatchSize": batch_size,
-        "BatchSum": batch_sum,
-        "BatchSquareSum": batch_square_sum,
-    }
-    attrs = {
-        "epsilon": epsilon,
-        "data_layout": data_layout,
-        "sync_stats": sync_stats,
-        "summary_decay_rate": summary_decay_rate,
-    }
-    if slot_dim > 0:
-        attrs["slot_dim"] = slot_dim
-    if enable_scale_and_shift:
-        attrs["enable_scale_and_shift"] = enable_scale_and_shift
-    if enable_scale_and_shift:
-        inputs["scale_w"] = scale_w
-        inputs["bias"] = bias
-    helper.append_op(
-        type="data_norm",
-        inputs=inputs,
-        outputs={
-            "Y": data_norm_out,
-            "Means": means,
-            "Scales": scales,
-            "BatchSize": batch_size,
-            "BatchSum": batch_sum,
-            "BatchSquareSum": batch_square_sum,
-        },
-        attrs=attrs,
-    )
-
-    return helper.append_activation(data_norm_out)
-
-
 @templatedoc()
 def layer_norm(
     input,
@@ -2812,116 +2606,6 @@ def layer_norm(
     )
 
     return helper.append_activation(layer_norm_out)
-
-
-@templatedoc()
-def group_norm(
-    input,
-    groups,
-    epsilon=1e-05,
-    param_attr=None,
-    bias_attr=None,
-    act=None,
-    data_layout='NCHW',
-    name=None,
-):
-    """
-    :api_attr: Static Graph
-
-    **Group Normalization Layer**
-
-    Refer to `Group Normalization <https://arxiv.org/abs/1803.08494>`_ .
-
-    Parameters:
-        input(Tensor): Tensor with dimension greater than 1, the data type is float32 or float64.
-        groups(int): The number of groups that divided from channels, the data type
-            is int32.
-        epsilon(float, optional): The small value added to the variance to prevent
-            division by zero, the data type is float32. Default: 1e-05.
-        param_attr(ParamAttr|bool, optional): ParamAttr object that specifies weight parameter
-            attribute. If a bool type, only False is supported, which means there is no weight parameter.
-            Default: None, the default weight parameter attribute is used. For more information, please
-            refer to :ref:`api_guide_ParamAttr` .
-        bias_attr(ParamAttr|bool, optional): ParamAttr object that specifies bias parameter
-            attribute. If a bool type, only False is supported, which means there is no bias parameter.
-            Default: None, the default bias parameter attribute is used. For more information, please
-            refer to :ref:`api_guide_ParamAttr` .
-        act(str, optional): Activation to be applied to the output of group normalization.
-        data_layout(str, optional): Specify the data format of the input, and the data format of the output
-            will be consistent with that of the input. An optional string from: `"NCHW"`, `"NHWC"`.
-            The default is `"NCHW"`. When it is `"NCHW"`, the data is stored in the order of:
-            `[batch_size, input_channels, *]`.
-        name (str, optional): The default value is None. Normally there is no need for user to set this
-            property. For more information, please refer to :ref:`api_guide_Name` .
-
-    Returns:
-        Tensor: A Tensor has same data type and data format with `input`.
-
-    Examples:
-       .. code-block:: python
-
-            import paddle
-            paddle.enable_static()
-
-            data = paddle.static.data(name='data', shape=[2, 8, 32, 32], dtype='float32')
-            x = paddle.static.nn.group_norm(input=data, groups=4)
-            print(x.shape) # [2, 8, 32, 32]
-    """
-    helper = LayerHelper('group_norm', **locals())
-    dtype = helper.input_dtype()
-    check_variable_and_dtype(
-        input, 'input', ['float32', 'float64'], 'group_norm'
-    )
-    # create intput and parameters
-    inputs = {'X': input}
-    input_shape = input.shape
-    if len(input_shape) < 2:
-        raise ValueError(
-            f"The dimensions of Op(fluid.layers.group_norm)'s input should be more than 1. But received {len(input_shape)}"
-        )
-    if data_layout != 'NCHW' and data_layout != 'NHWC':
-        raise ValueError(
-            "Param(data_layout) of Op(fluid.layers.group_norm) got wrong value: received "
-            + data_layout
-            + " but only NCHW or NHWC supported."
-        )
-    channel_num = input_shape[1] if data_layout == 'NCHW' else input_shape[-1]
-    param_shape = [channel_num]
-    if param_attr:
-        scale = helper.create_parameter(
-            attr=helper.param_attr,
-            shape=param_shape,
-            dtype=dtype,
-            default_initializer=Constant(1.0),
-        )
-        inputs['Scale'] = scale
-    if bias_attr:
-        bias = helper.create_parameter(
-            attr=helper.bias_attr, shape=param_shape, dtype=dtype, is_bias=True
-        )
-        inputs['Bias'] = bias
-
-    # create output
-    mean_out = helper.create_variable(dtype=dtype, stop_gradient=True)
-    variance_out = helper.create_variable(dtype=dtype, stop_gradient=True)
-    group_norm_out = helper.create_variable(dtype=dtype)
-
-    helper.append_op(
-        type="group_norm",
-        inputs=inputs,
-        outputs={
-            "Y": group_norm_out,
-            "Mean": mean_out,
-            "Variance": variance_out,
-        },
-        attrs={
-            "epsilon": epsilon,
-            "groups": groups,
-            "data_layout": data_layout,
-        },
-    )
-
-    return helper.append_activation(group_norm_out)
 
 
 @templatedoc()
