+++ python/paddle/fluid/tests/unittests/ipu/test_groupnorm_op_ipu.py
@@ -63,7 +63,7 @@ class TestBase(IPUOpTest):
             )
             scale = paddle.ParamAttr(trainable=True)
             bias = paddle.ParamAttr(trainable=True)
-            out = paddle.fluid.layers.nn.group_norm(
+            out = paddle.static.nn.group_norm(
                 conv1, param_attr=scale, bias_attr=bias, **self.attrs
             )
             loss = paddle.mean(out)
@@ -71,7 +71,7 @@ class TestBase(IPUOpTest):
             adam.minimize(loss)
             self.fetch_list = [loss.name]
         else:
-            out = paddle.fluid.layers.nn.group_norm(
+            out = paddle.static.nn.group_norm(
                 x, param_attr=True, bias_attr=True, **self.attrs
             )
             self.fetch_list = [out.name]
