+++ python/paddle/fluid/tests/unittests/test_data_norm_op.py
@@ -15,6 +15,7 @@
 
 import unittest
 import numpy as np
+import paddle
 import paddle.fluid.core as core
 from paddle.fluid.op import Operator
 import paddle.fluid as fluid
@@ -517,7 +518,7 @@ class TestDataNormOpErrorr(unittest.TestCase):
         with program_guard(Program(), Program()):
             x2 = fluid.layers.data(name='x2', shape=[3, 4], dtype="int32")
             # self.assertRaises(TypeError, fluid.data_norm, x2)
-            fluid.layers.data_norm(
+            paddle.static.nn.data_norm(
                 input=x2, param_attr={}, enable_scale_and_shift=True
             )
