+++ python/paddle/fluid/tests/unittests/test_dist_fleet_ps2.py
@@ -89,7 +89,7 @@ class TestPSPassWithBow(unittest.TestCase):
         # vsum
         q_sum = fluid.layers.sequence_pool(input=q_emb, pool_type='sum')
         q_ss = paddle.nn.functional.softsign(q_sum)
-        q_ss = fluid.layers.data_norm(input=q_ss)
+        q_ss = paddle.static.nn.data_norm(input=q_ss)
         # fc layer after conv
         q_fc = fluid.layers.fc(
             input=q_ss,
