+++ python/paddle/fluid/tests/unittests/test_fleet.py
@@ -32,6 +32,7 @@ class TestFleet1(unittest.TestCase):
 
     def test_pslib_1(self):
         """Test cases for pslib."""
+        import paddle
         import paddle.fluid as fluid
         from paddle.fluid.incubate.fleet.parameter_server.pslib import fleet
         from paddle.fluid.incubate.fleet.base.role_maker import GeneralRoleMaker
@@ -66,7 +67,9 @@ class TestFleet1(unittest.TestCase):
                 param_attr=fluid.ParamAttr(name="embedding"),
             )
             bow = fluid.layers.sequence_pool(input=emb, pool_type='sum')
-            bow = fluid.layers.data_norm(input=bow, epsilon=1e-4, name="norm")
+            bow = paddle.static.nn.data_norm(
+                input=bow, epsilon=1e-4, name="norm"
+            )
             fc = fluid.layers.fc(input=bow, size=1, act=None)
             label = fluid.layers.data(
                 name="click",
